UBUNTU: [Packaging] simplify revision/release
Use "dpkg-parsepackage -S version" to get the version, and use Make's
built-in functions to split it into the upsteam-version and the
debian revision. This is more readable and works more efficiently.

BTW, GNU Make >= 3.81 supports $(lastword x) as a shorthand of
$(word $(words x), x).

Drop "revisions" and "prev_revisions", which are unused anywhere
except "printenv".

I did not get the use of "?=" for revision and prev_revison, but
I am keeping them in case I am missing something here.

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>
Acked-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>