UBUNTU: SAUCE: sched-ext: taint kernel when a custom scheduler is loaded
BugLink: https://bugs.launchpad.net/bugs/2045503

In case of potential bug reports about performance regressions we want
to be able to easily determine if a custom scheduler has ever been used
(that could be the reason of the regression).

For this reason set the TAINT_OOT_MODULE taint flag in the kernel:
loading a custom scheduler should be considered exactly as loading an
out-of-tree kernel module.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>