ci: add github workflow to test the sched-ext kernel
Add a github action to test the sched-ext kernel with all the shipped
schedulers.

The test uses a similar approach to the scx workflow [1], using
virtme-ng to run each scheduler inside a sched-ext enabled kernel for a
certain amount of time (30 sec) and checking for potential stall, oops
or bug conditions.

In this case we can use `virtme-ng --build` to build a kernel with bare
minimum support to run inside virtme-ng itself, instead of generating a
fully featured kernel, to expedite the testing process.

The mandatory .config options required by sched-ext are stored in
`.github/workflows/sched-ext.config` and they are passed to virtme-ng
via the `--config` option.

The test itself is defined in `.github/workflows/run-schedulers`: the
script looks for all the binaries in `tools/sched_ext/build/bin` and
runs each one in a separate virtme-ng instance, to ensure that each run
does not impact the others.

[1] https://github.com/sched-ext/scx/blob/main/.github/workflows/build-scheds.yml

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>